pyproject.toml
[project]
name = "cardagon-server"
version = "1.0.0"
dependencies = [
  "fastapi[standard]==0.120.4",
  "sqlalchemy==2.0.44",
  "pydantic-settings==2.11.0",
  "passlib==1.7.4",
  "pyjwt==2.10.1",
  "psycopg2-binary==2.9.11",
  "alembic==1.17.1",
  "argon2_cffi==25.1.0",
]
requires-python = "==3.13.9"


[tool.black]
line-length = 88

[tool.alembic]
script_location = "%(here)s/alembic"
prepend_sys_path = ["."]
